+++ crates/bevy-inspector-egui/examples/basic/inspector_options.rs
@@ -91,7 +91,7 @@ fn ui_example(world: &mut World) {
         .clone();
 
     egui::Window::new("UI").show(egui_context.get_mut(), |ui| {
-        egui::ScrollArea::vertical().show(ui, |ui| {
+        egui::ScrollArea::both().show(ui, |ui| {
             bevy_inspector_egui::bevy_inspector::ui_for_resource::<UiData>(world, ui);
         });
     });